Allow arbitrary raster params to be passed through to the request
This makes it easier to prototype or stage new features without
hacking the client locally.

We lose a bit of type checking on the client side, but all
params are also checked on the server anyway. Users get in any
case a helpful message pointing out, e.g., typos in argument names.