Change line continuation in docstring examples
With backslash line continuations, the lines are merged into
a single line in the documentation output on readthedocs. Using
repl continuation is supported by doctest and is much more
readable in the docs.

Caveat is that the examples are less copyable, I could be
convinced the previous style is better.